Fix config register byte order (#19)
The driver currently assumes byte 1 of the config register is the LSB
and byte 2 is the MSB. This PR changes it to the correct order (byte 1
is MSB, byte 2 is LSB).

This was discovered after writing `Config::default()` to a TMP108
device. The device was not returning fresh temperature samples during
periodic polling. Likely the device was in shutdown mode since the
current default value written would be `0x1022` which would cause M1 and
M0 bits to be set to `0b00` (shutdown mode).

This fix thus changes the default value to `0x2210` which makes the
device operate as expected.