Update to embedded-sensors v0.3.0 (#21)
Updates to `embedded-sensors` v0.3.0, introducing breaking changes to
the API requiring update of `tmp108` version from v0.2.0 to v0.3.0.

MSRV changed to 1.84 to be able to use f32::abs from core (which was
moved from std). Alternatives are to use silent integer truncation
instead of checking for float equality within an epsilon, but that might
result in behavior caller does not expect.